// src/web/src/api/generated/models/HealthContextResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type HealthContextResponse = {
    id: string;
    contextType: string;
    bodyLocation: string | null;
    startedOn: string | null;
    endedOn: string | null;
    status: string;
    description: string;
    updatedAt: string;
};