// src/web/src/api/generated/models/SaveHealthContextRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type SaveHealthContextRequest = {
    contextType: string;
    bodyLocation: string | null;
    startedOn: string | null;
    endedOn: string | null;
    status: string;
    description: string;
};