// src/web/src/api/generated/models/UpdateAthleteProfileRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type UpdateAthleteProfileRequest = {
    displayName: string;
    birthDate: string | null;
    heightCm: number | string | null;
    weightKg: number | string | null;
    sex: string;
    timezoneName: string;
    locale: string;
    unitSystem: string;
};